Move monadic coroutines to their own class. Add AtomicReference() to fetch returned monad
