+++ kategory-core/src/main/kotlin/kategory/typeclasses/Monad.kt
@@ -1,13 +1,6 @@
 package kategory
 
-import java.io.Serializable
-import kotlin.coroutines.experimental.Continuation
-import kotlin.coroutines.experimental.CoroutineContext
-import kotlin.coroutines.experimental.EmptyCoroutineContext
-import kotlin.coroutines.experimental.RestrictsSuspension
-import kotlin.coroutines.experimental.intrinsics.COROUTINE_SUSPENDED
-import kotlin.coroutines.experimental.intrinsics.suspendCoroutineOrReturn
-import kotlin.coroutines.experimental.startCoroutine
+import kotlin.coroutines.experimental.*
 
 interface Monad<F> : Applicative<F>, Typeclass {
 
@@ -38,84 +31,15 @@ inline fun <reified F, A, B> HK<F, A>.flatMap(FT: Monad<F> = monad(), noinline f
 
 inline fun <reified F, A> HK<F, HK<F, A>>.flatten(FT: Monad<F> = monad()): HK<F, A> = FT.flatten(this)
 
-@RestrictsSuspension
-open class MonadContinuation<F, A>(M: Monad<F>, override val context: CoroutineContext = EmptyCoroutineContext) :
-        Serializable, Continuation<HK<F, A>>, Monad<F> by M {
-
-    override fun resume(value: HK<F, A>) {
-        returnedMonad = value
-    }
-
-    override fun resumeWithException(exception: Throwable) {
-        throw exception
-    }
-
-    protected lateinit var returnedMonad: HK<F, A>
-
-    internal fun returnedMonad(): HK<F, A> = returnedMonad
-
-    suspend fun <B> HK<F, B>.bind(): B = bind { this }
-
-    suspend fun <B> bind(m: () -> HK<F, B>): B = suspendCoroutineOrReturn { c ->
-        val labelHere = c.stackLabels // save the whole coroutine stack labels
-        returnedMonad = flatMap(m(), { x: B ->
-            c.stackLabels = labelHere
-            c.resume(x)
-            returnedMonad
-        })
-        COROUTINE_SUSPENDED
-    }
-
-    infix fun <B> yields(b: B): HK<F, B> = yields { b }
-
-    infix fun <B> yields(b: () -> B): HK<F, B> = pure(b())
-}
-
 /**
  * Entry point for monad bindings which enables for comprehension. The underlying impl is based on coroutines.
  * A coroutine is initiated and inside `MonadContinuation` suspended yielding to `flatMap` once all the flatMap binds are completed
  * the underlying monad is returned from the act of executing the coroutine
  */
-fun <F, B> Monad<F>.binding(c: suspend MonadContinuation<F, *>.() -> HK<F, B>): HK<F, B> {
-    val continuation = MonadContinuation<F, B>(this)
+fun <F, B> Monad<F>.binding(c: suspend MonadCoroutines<F, *>.() -> HK<F, B>): HK<F, B> {
+    val continuation = MonadCoroutines<F, B>(this)
     c.startCoroutine(continuation, continuation)
     return continuation.returnedMonad()
-}
-
-@RestrictsSuspension
-open class StackSafeMonadContinuation<F, A>(M: Monad<F>, override val context: CoroutineContext = EmptyCoroutineContext) :
-        Serializable, Continuation<Free<F, A>>, Monad<F> by M {
-
-    override fun resume(value: Free<F, A>) {
-        returnedMonad = value
-    }
-
-    override fun resumeWithException(exception: Throwable) {
-        throw exception
-    }
-
-    protected lateinit var returnedMonad: Free<F, A>
-
-    internal fun returnedMonad(): Free<F, A> = returnedMonad
-
-    suspend fun <B> HK<F, B>.bind(): B = bind { Free.liftF(this) }
-
-    suspend fun <B> Free<F, B>.bind(): B = bind { this }
-
-    suspend fun <B> bind(m: () -> Free<F, B>): B = suspendCoroutineOrReturn { c ->
-        val labelHere = c.stackLabels // save the whole coroutine stack labels
-        val freeResult = m()
-        returnedMonad = freeResult.flatMap { z ->
-            c.stackLabels = labelHere
-            c.resume(z)
-            returnedMonad
-        }
-        COROUTINE_SUSPENDED
-    }
-
-    infix fun <B> yields(b: B): Free<F, B> = yields { b }
-
-    infix fun <B> yields(b: () -> B): Free<F, B> = Free.liftF(pure(b()))
 }
 
 /**
